Code refactor
Refactored the code to make it easier to understand.
Added a flag to just print the notes of a given scale and exit
Added melodic and harmonic minor scales.
Changed the default minimum frames per note from 7 to 8.
Cut off the lower and upper octave from Square1 and Square2
respectively. This is mostly a test to see if this improves output.